+++ contracts/token/ERC20/ERC20.sol
@@ -2,7 +2,8 @@
 
 pragma solidity ^0.8.0;
 
-import "./IERC20Extended.sol";
+import "./IERC20.sol";
+import "./extensions/IERC20Metadata.sol";
 import "../../utils/Context.sol";
 
 /**
@@ -29,7 +30,7 @@ import "../../utils/Context.sol";
  * functions have been added to mitigate the well-known issues around setting
  * allowances. See {IERC20-approve}.
  */
-contract ERC20 is Context, IERC20Extended {
+contract ERC20 is Context, IERC20, IERC20Metadata {
     mapping (address => uint256) private _balances;
 
     mapping (address => mapping (address => uint256)) private _allowances;