+++ contracts/token/ERC20/extensions/IERC20Metadata.sol
@@ -2,12 +2,12 @@
 
 pragma solidity ^0.8.0;
 
-import "./IERC20.sol";
+import "../IERC20.sol";
 
 /**
- * @dev Interface of the extended ERC20 standard.
+ * @dev Interface for the optional metadata functions from the ERC20 standard.
  */
-interface IERC20Extended is IERC20 {
+interface IERC20Metadata is IERC20 {
     /**
      * @dev Returns the name of the token.
      */